requirements.txt
-e .
black
isort
numpy
pandas
scikit-learn
openpyxl
psycopg2-binary
pytest
pylint
pyyaml
sqlalchemy
pre-commit